Move grunt config to config folder
