+++ config/grunt/task-config/clean.js
